[skip ci] 1.18.2 update

+++ src/main/java/io/github/darkkronicle/advancedchatcore/chat/AdvancedChatScreen.java
@@ -247,7 +247,7 @@ public class AdvancedChatScreen extends GuiBase {
         }
 
         // Send to hud to scroll
-        client.inGameHud.getChatHud().scroll(amount);
+        client.inGameHud.getChatHud().scroll((int) amount);
         return true;
     }
